+++ src/engine/systems/index.ts
@@ -1,3 +1,4 @@
 export { ScriptSystem } from './ScriptSystem.ts'
 export { AnimationSystem } from './AnimationSystem.ts'
+export { AudioSystem } from './AudioSystem.ts'
 export { UISystem } from './UISystem.ts'